DSP-8 init firebase
