+++ firebase/.firebaserc
@@ -1,9 +1,5 @@
 {
   "projects": {
-    "default": "eba-dsp-map-dev",
-    "prod": "eba-dsp-map-prod"
-  },
-  "targets": {},
-  "etags": {},
-  "dataconnectEmulatorConfig": {}
+    "default": "eba-dsp-map-dev"
+  }
 }
