--- firebase/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Firebase Example</title>
-    <script type="module" src="firebase.js"></script>
-</head>
-
-<body>
-    <h1>Hello, from Firebase!</h1>
-</body>
-
-</html>
