+++ firebase/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "firebase",
+  "version": "1.0.0",
+  "main": "fs_emulator_connect.js",
+  "type": "module",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "description": ""
+}
